fixed concurrency problem in chain commit, fixes #463
